chore: split matterbridge-home-assistant into multiple libraries
BREAKING CHANGE: In version 1.x `matterbridge` was not listed as a `dependencies` or `peerDependency`. This has been changed in version 2.0.0. It is now listed as a peer dependency.
When installed in a local `package.json` file, this is not a problem. But since `matterbridge` installs all its plugins globally, this will lead to an error running `matterbridge` with `matterbridge-home-assistant`, bacause of npm's `new` (>= 7) strategy for peer-dependencies.
To solve this, you need to enable [legacy-peer-deps](https://docs.npmjs.com/cli/v10/using-npm/config#legacy-peer-deps) in your npm config (`npm config set legacy-peer-deps true`). In the pre-built docker image and the native Home Assistant Addon, this is already configured.
**This change is only "breaking" if you installed this project by hand using npm or within matterbridge.**